Create README.md
This Matlab toolbox allows the study of the behavior of a complex system by introducing a new approach capable of tracking changes in the information stored in the system over time. This can be achieved through the recursive least squares (RLS) algorithm and Yule-Walker equations, assuming Gaussianity. The capabilities of this approach have been used to study heart-brain interactions